Fixes for older Python versions.
