+++ pyVHDLModel/Concurrent.py
@@ -163,14 +163,14 @@ class ConfigurationInstantiation(Instantiation):
 
 @export
 class ProcessStatement(ConcurrentStatement, SequentialDeclarations, SequentialStatements, DocumentedEntityMixin):
-	_sensitivityList: List[Name]  # TODO: implement a SignalSymbol
+	_sensitivityList: List['Name']  # TODO: implement a SignalSymbol
 
 	def __init__(
 		self,
 		label: str = None,
 		declaredItems: Iterable = None,
 		statements: Iterable[SequentialStatement] = None,
-		sensitivityList: Iterable[Name] = None,
+		sensitivityList: Iterable['Name'] = None,
 		documentation: str = None
 	):
 		super().__init__(label)
@@ -187,13 +187,13 @@ class ProcessStatement(ConcurrentStatement, SequentialDeclarations, SequentialSt
 				# signalSymbol._parent = self  # FIXME: currently str are provided
 
 	@property
-	def SensitivityList(self) -> List[Name]:
+	def SensitivityList(self) -> List['Name']:
 		return self._sensitivityList
 
 
 @export
 class ConcurrentProcedureCall(ConcurrentStatement, ProcedureCall):
-	def __init__(self, label: str, procedureName: Name, parameterMappings: Iterable[ParameterAssociationItem] = None):
+	def __init__(self, label: str, procedureName: 'Name', parameterMappings: Iterable[ParameterAssociationItem] = None):
 		super().__init__(label)
 		ProcedureCall.__init__(self, procedureName, parameterMappings)
 
@@ -473,7 +473,7 @@ class ForGenerateStatement(GenerateStatement, ConcurrentDeclarations, Concurrent
 
 @export
 class ConcurrentSignalAssignment(ConcurrentStatement, SignalAssignment):
-	def __init__(self, label: str, target: Name):
+	def __init__(self, label: str, target: 'Name'):
 		super().__init__(label)
 		SignalAssignment.__init__(self, target)
 
@@ -482,7 +482,7 @@ class ConcurrentSignalAssignment(ConcurrentStatement, SignalAssignment):
 class ConcurrentSimpleSignalAssignment(ConcurrentSignalAssignment):
 	_waveform: List[WaveformElement]
 
-	def __init__(self, label: str, target: Name, waveform: Iterable[WaveformElement]):
+	def __init__(self, label: str, target: 'Name', waveform: Iterable[WaveformElement]):
 		super().__init__(label, target)
 
 		# TODO: extract to mixin
@@ -499,13 +499,13 @@ class ConcurrentSimpleSignalAssignment(ConcurrentSignalAssignment):
 
 @export
 class ConcurrentSelectedSignalAssignment(ConcurrentSignalAssignment):
-	def __init__(self, label: str, target: Name, expression: ExpressionUnion):
+	def __init__(self, label: str, target: 'Name', expression: ExpressionUnion):
 		super().__init__(label, target)
 
 
 @export
 class ConcurrentConditionalSignalAssignment(ConcurrentSignalAssignment):
-	def __init__(self, label: str, target: Name, expression: ExpressionUnion):
+	def __init__(self, label: str, target: 'Name', expression: ExpressionUnion):
 		super().__init__(label, target)
